chore: Split features list
Also bullets instead of pipes in notification text, new NMEA icon
