+++ GPSTest/src/main/res/drawable/ic_nmea.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="21.037403dp"
+    android:height="10.00968dp"
+    android:viewportWidth="21.037403"
+    android:viewportHeight="10.00968">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="m5.1935,6.1805q0,0.801 -0.6253,1.3333 -0.6253,0.5271 -1.6381,0.6253v1.8707L2.3203,10.0097L2.3203,8.1648q-0.6821,-0.005 -1.2816,-0.1292 -0.5994,-0.1292 -1.0387,-0.3307v-1.0232h0.0827q0.0982,0.0724 0.3514,0.2119 0.2532,0.1344 0.4909,0.2222 0.2687,0.0982 0.6253,0.186 0.3617,0.0827 0.77,0.0982v-2.2376q-0.2067,-0.0413 -0.3824,-0.0775 -0.1757,-0.0413 -0.3256,-0.0827 -0.8423,-0.2119 -1.2092,-0.6356 -0.3669,-0.4289 -0.3669,-1.0542 0,-0.7648 0.5994,-1.2919 0.6046,-0.5271 1.6846,-0.615v-1.4056h0.6098v1.3953q0.5219,0.0103 1.0697,0.124 0.5478,0.1137 0.9198,0.2636v1.0129h-0.0724q-0.3876,-0.2377 -0.8113,-0.4186 -0.4186,-0.186 -1.1059,-0.2325v2.2273q0.155,0.0258 0.3359,0.0724 0.1809,0.0413 0.3152,0.0672 0.77,0.1654 1.1886,0.5684 0.4237,0.4031 0.4237,1.1007zM2.3203,4.284v-2.1342q-0.5529,0.0413 -0.9302,0.3049 -0.3772,0.2584 -0.3772,0.7235 0,0.4703 0.2791,0.708 0.2791,0.2377 1.0284,0.3979zM4.2168,6.3148q0,-0.4858 -0.3049,-0.708 -0.2997,-0.2274 -0.9818,-0.3566v2.1394q0.6201,-0.062 0.9508,-0.3152 0.3359,-0.2532 0.3359,-0.7596z"
+      android:strokeWidth="0.264583"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="m13.379,7.5757q-0.6305,0.2894 -1.3798,0.5064 -0.7441,0.2119 -1.4418,0.2119 -0.8992,0 -1.6485,-0.2481 -0.7493,-0.248 -1.2764,-0.7441 -0.5323,-0.5013 -0.8217,-1.2506 -0.2894,-0.7545 -0.2894,-1.7622 0,-1.8448 1.0749,-2.9094 1.08,-1.0697 2.9611,-1.0697 0.6563,0 1.3384,0.1602 0.6873,0.155 1.4779,0.5323v1.2144h-0.093q-0.1602,-0.124 -0.4651,-0.3256 -0.3049,-0.2015 -0.5994,-0.3359 -0.3566,-0.1602 -0.8113,-0.2636 -0.4496,-0.1085 -1.0232,-0.1085 -1.2919,0 -2.0464,0.832 -0.7493,0.8268 -0.7493,2.2427 0,1.4935 0.7855,2.3254 0.7855,0.8268 2.1394,0.8268 0.4961,0 0.987,-0.0982 0.4961,-0.0982 0.8682,-0.2532v-1.8862h-2.0619v-0.8992h3.0747z"
+      android:strokeWidth="0.264583"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="m21.0374,8.1442h-1.2661l-3.6483,-6.8833v6.8833L15.167,8.1442v-7.6946h1.5865l3.3279,6.2838v-6.2838h0.956z"
+      android:strokeWidth="0.264583"/>
+</vector>
