+++ GPSTest/src/main/res/values/strings.xml
@@ -134,7 +134,7 @@
     <string name="error_time_message">The device GNSS time of:\n\n%1$s\n\n…is more than %2$d days different from your system clock. This can be caused by the GPS week rollover issue - ask your device manufacturer for details. See https://bit.ly/gps-week-rollover for more info.</string>
     <string name="unknown">Unknown</string>
     <string name="cancel">Cancel</string>
-    <string name="notification_title">%1$d/%2$d sats | %3$d/%4$d signals</string>
+    <string name="notification_title">%1$d/%2$d sats &#8226; %3$d/%4$d signals</string>
 
     <string name="gps_content_description">GPS</string>
     <string name="glonass_content_description">GLONASS</string>
@@ -215,7 +215,10 @@
 
     <string name="dashboard_supported_sbas">Supported SBAS</string>
 
-    <string name="dashboard_feature_support">Feature support</string>
+    <string name="dashboard_feature_accuracy">Features that improve accuracy</string>
+    <string name="dashboard_feature_info">Features for GNSS debug</string>
+    <string name="dashboard_feature_assist_data">Features that modify assist data</string>
+
     <string name="dashboard_supported">Supported</string>
     <string name="dashboard_not_supported">Not supported</string>
